fixed some nwffacim study group formatting errors

+++ work/search/nwffacim/acim/prep.sh
@@ -1,8 +1,16 @@
-rm *.json
 
-for i in `cat contents`; do
+list="contents"
+
+if [ "$1" != "" ]; then
+  list=$1
+else
+  rm *.json
+fi
+
+for i in `cat $list`; do
 book=${i:0:4}
 unit=${i:5}
-#echo "i=$i, book=$book, unit=$unit"
+#echo "list=$list, i=$i, book=$book, unit=$unit"
 prep -s nwffacim -b $book $unit
 done
+
